remove "required=False" for options with "default" settings
The default "required" value for AnsibleModule's argument_spec is
"False". There is no need to specify it in our code.

Also, the Ansible documentation explicitly says that required "should
not be used in combination with default".

Remove the "required=False" settings. This simplifies the code and makes
it easier to read.